% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_fmp_fx}
\alias{get_fmp_fx}
\title{Get FX prices from FMP}
\usage{
get_fmp_fx(
  pair = "EURUSD",
  frequency = "1min",
  from = "2022-03-02",
  to = from,
  apikey = read_file(".apikey")
)
}
\arguments{
\item{pair}{e.g. `"EURUSD"`}

\item{frequency}{e.g. `"1min"`}

\item{from}{date}

\item{to}{date}

\item{apikey}{API key for https://site.financialmodelingprep.com/developer/docs}
}
\value{
price data
}
\description{
Get FX prices from FMP
}
