% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{net_position}
\alias{net_position}
\title{Compute net position and profit}
\usage{
net_position(trades_l, lot_size = 1e+05, col_buysell = "Action")
}
\arguments{
\item{trades_l}{long trade data, normally created with `trades_long()`}

\item{lot_size}{standard lot size is 100000}

\item{col_buysell}{default `"Action"`}
}
\value{
trades_l but with position and profit columns
}
\description{
Compute net position and profit
}
\examples{
# will follow
}
