% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trades_long}
\alias{trades_long}
\title{Get longer format trades}
\usage{
trades_long(
  trades,
  col_opendate = "Open Date",
  col_closedate = "Close Date",
  col_openprice = "Open Price",
  col_closeprice = "Close Price",
  cols_to_keep = c("Lots", "Action", "Profit")
)
}
\arguments{
\item{trades}{trade data}

\item{col_opendate}{default `"Open Date"`}

\item{col_closedate}{default `"Close Date"`}

\item{col_openprice}{default `"Open Price"`}

\item{col_closeprice}{default `"Close Price"`}

\item{cols_to_keep}{columns to keep, default `c("Lots", "Action", "Profit")`}
}
\value{
long trade data
}
\description{
Get longer format trades
}
\examples{
trades |>
  filter(Action \%in\% c("Buy", "Sell")) |>
  trades_long()
}
